requirements.txt
Babel==2.7.0
bcrypt==3.1.7
blinker==1.4
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
dominate==2.4.0
fabric==2.5.0
Flask==1.1.1
Flask-Assets==0.12
Flask-BabelEx==0.9.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
flask-nav==0.6
Flask-Principal==0.4.0
Flask-Security-Too==3.3.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
googlemaps==3.1.4
idna==2.8
invoke==1.3.0
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mysqlclient==1.4.6
numpy==1.17.4
paramiko==2.7.0
passlib==1.7.2
pycparser==2.19
PyNaCl==1.3.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.13.0
speaklater==1.3
SQLAlchemy==1.3.11
sqlalchemy-datatables==2.0.1
unicodecsv==0.14.1
urllib3==1.25.7
visitor==0.1.3
webassets==0.12.1
Werkzeug==0.16.0
WTForms==2.2.1
